# Detailed/03-OpenAPI/med-rms-openapi.yaml
...
openapi: 3.0.3
info:
  title: Med-RMS 医疗器械需求管理系统 API
  description: 'Med-RMS (Medical Device Requirement Management System) RESTful API 规范。

    涵盖需求管理、追溯管理、变更管理、合规管理、电子签名、风险管理、项目管理、报表仪表盘及系统管理模块。

    '
  version: 1.1.0
  contact:
    name: Med-RMS 开发团队
  license:
    name: Proprietary
servers:
- url: http://localhost:8080/api/v1
  description: 开发环境
...
security:
- BearerAuth: []
paths:
...
  /requirements/{id}:
    get:
      operationId: getRequirement
      tags:
      - 需求管理
      summary: 查询需求详情
      parameters:
      - $ref: '#/components/parameters/RequirementIdParam'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/RequirementDetailVO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      x-required-permission: req:list
    put:
      operationId: updateRequirement
      tags:
      - 需求管理
      summary: 更新需求
      parameters:
      - $ref: '#/components/parameters/RequirementIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequirementUpdateDTO'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/RequirementVO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
      x-required-permission: req:update
    delete:
      operationId: deleteRequirement
      tags:
      - 需求管理
      summary: 删除需求（仅Draft状态）
      parameters:
      - $ref: '#/components/parameters/RequirementIdParam'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
      x-required-permission: req:delete
  /requirements/{id}/status:
    patch:
      operationId: changeRequirementStatus
      tags:
      - 需求管理
      summary: 变更需求状态
      parameters:
      - $ref: '#/components/parameters/RequirementIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusChangeDTO'
      responses:
        '200':
          description: 状态变更成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/RequirementVO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
      x-required-permission: req:update
...
  /requirements/{id}/versions:
    post:
      operationId: createRequirementVersion
      tags:
      - 需求管理
      summary: 创建需求版本
      parameters:
      - $ref: '#/components/parameters/RequirementIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionCreateDTO'
      responses:
        '201':
          description: 版本创建成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/RequirementVersionVO'
        '400': null
        $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      x-required-permission: req:create
    get:
      operationId: getRequirementVersions
      tags:
      - 需求管理
      summary: 查询版本列表
      parameters:
      - $ref: '#/components/parameters/RequirementIdParam'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      type: array
                      items: null
          $ref: '#/components/schemas/RequirementVersionVO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      x-required-permission: req:list
...
  /requirements/{id}/review:
    post:
      operationId: createReview
      tags:
      - 需求管理
      summary: 发起评审
      parameters:
      - $ref: '#/components/parameters/RequirementIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewCreateDTO'
      responses:
        '201':
          description: 评审创建成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/ReviewVO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
      x-required-permission: req:review
  /requirements/{id}/reviews:
    get:
      operationId: getReviewList
      tags:
      - 需求管理
      summary: 查询评审记录
      parameters:
      - $ref: '#/components/parameters/RequirementIdParam'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties: null
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReviewVO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      x-required-permission: req:list
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    RequirementIdParam:
      name: id
      in: path
      required: true
      description: 需求ID
      schema:
        type: string
        format: int64
  responses:
    BadRequest:
      description: 请求参数校验失败
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                data:
                  $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 未认证（Token缺失或过期）
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                data:
                  $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: 无权限访问
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                data:
                  $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                data:
                  $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: 资源冲突
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                data:
                  $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: 系统内部错误
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                data:
                  $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          description: 响应状态码
          example: 200
        message:
          type: string
          description: 响应消息
          example: success
        timestamp:
          type: string
          format: date-time
          description: 响应时间戳
...
    ErrorResponse:
      type: object
      required:
      - errorCode
      - errorMessage
      properties:
        errorCode:
          type: string
          description: 错误码
          example: SY0100
        errorMessage:
          type: string
          description: 错误消息
          example: 请求参数校验失败
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
          description: 字段级错误详情
...
    ReqStatus:
      type: string
      enum:
      - Draft
      - PendingDecompose
      - Decomposed
      - Submitted
      - InReview
      - Approved
      - Rejected
      - PendingVerify
      - Implemented
      - Verified
      - Baseline
      - Changed
      - Closed
      - Retired
      description: 需求状态
    ReqType:
      type: string
      enum:
      - URS
      - PRS
      - SRS
      - DRS
      description: 需求类型（用户/产品/软件/设计）
    ReqPriority:
      type: string
      enum:
      - MUST
      - SHOULD
      - COULD
      - WONT
      description: 需求优先级（MoSCoW）
...
    ReviewResult:
      type: string
      enum:
      - Approved
      - Rejected
      - Conditional
      description: 评审结果
...
    RequirementUpdateDTO:
      type: object
      properties:
        title:
          type: string
          description: 需求标题
          maxLength: 200
        description:
          type: string
          description: 需求详细描述
        requirementType:
          $ref: '#/components/schemas/ReqType'
        priority:
          $ref: '#/components/schemas/ReqPriority'
        riskLevel:
          type: string
          enum:
          - HIGH
          - MEDIUM
          - LOW
          description: 风险等级
        safetyClass:
          type: string
          enum:
          - A
          - B
          - C
          description: 软件安全分类
        requirementCategory:
          type: string
          enum:
          - SOFTWARE
          - HARDWARE
          - BOTH
          description: 需求分类
        version:
          type: integer
          description: 乐观锁版本号
    RequirementVO:
      type: object
      properties:
        id:
          type: string
          format: int64
          description: 需求ID
        requirementNo:
          type: string
          description: 需求编号（自动生成，如URS-P001-001）
        projectId:
          type: string
          format: int64
          description: 项目ID
        projectName:
          type: string
          description: 项目名称
        parentId:
          type: string
          format: int64
          description: 父需求ID
        title:
          type: string
          description: 需求标题
        requirementType:
          $ref: '#/components/schemas/ReqType'
        status:
          $ref: '#/components/schemas/ReqStatus'
        priority:
          $ref: '#/components/schemas/ReqPriority'
        riskLevel:
          type: string
          enum:
          - HIGH
          - MEDIUM
          - LOW
          description: 风险等级
        safetyClass:
          type: string
          enum:
          - A
          - B
          - C
          description: 软件安全分类（IEC 62304）
        requirementCategory:
          type: string
          enum:
          - SOFTWARE
          - HARDWARE
          - BOTH
          description: 需求分类
        baselineId:
          type: string
          format: int64
          description: 关联基线ID
        version:
          type: integer
          description: 乐观锁版本号
        isDeleted:
          type: boolean
          description: 软删除标记
        createdBy:
          type: string
          description: 创建人
        createdAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          description: 最后修改人
        updatedAt:
          type: string
          format: date-time
    RequirementDetailVO:
      type: object
      allOf:
      - $ref: '#/components/schemas/RequirementVO'
      - type: object
        properties:
          description:
            type: string
            description: 需求详细描述
          source:
            type: string
            description: 需求来源
          childrenCount:
            type: integer
            description: 子需求数量
          traceCount:
            type: integer
            description: 关联追溯数量
...
    StatusChangeDTO:
      type: object
      required:
      - targetStatus
      - reason
      properties:
        targetStatus:
          type: string
          description: 目标状态
        reason:
          type: string
          description: 变更原因
          maxLength: 500
        remark:
          type: string
          description: 备注
    VersionCreateDTO:
      type: object
      required:
      - changeDescription
      properties:
        changeDescription:
          type: string
          description: 版本变更说明
          maxLength: 500
    RequirementVersionVO:
      type: object
      properties:
        id:
          type: string
          format: int64
        requirementId:
          type: string
          format: int64
          description: 需求ID
        version:
          type: integer
          description: 版本号
        changeDescription:
          type: string
          description: 变更说明
        snapshot:
          type: object
          description: 需求快照
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
    ReviewCreateDTO:
      type: object
      required:
      - reviewerIds
      properties:
        reviewerIds:
          type: array
          items:
            type: string
            format: int64
          description: 评审人ID列表
        reviewType:
          type: string
          enum:
          - PeerReview
          - FormalReview
          description: 评审类型
        comment:
          type: string
          description: 评审说明
    ReviewVO:
      type: object
      properties:
        id:
          type: string
          format: int64
        requirementId:
          type: string
          format: int64
        reviewType:
          type: string
          description: 评审类型
        status:
          type: string
          enum:
          - Pending
          - InProgress
          - Completed
          description: 评审状态
        result:
          $ref: '#/components/schemas/ReviewResult'
        reviewers:
          type: array
          items:
            type: object
            properties:
              userId:
                type: string
                format: int64
              username:
                type: string
              result:
                $ref: '#/components/schemas/ReviewResult'
              comment:
                type: string
              reviewedAt:
                type: string
                format: date-time
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
...